pyproject.toml
[project]
name = "team13-aws-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.24",
    "django>=5.2.6",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.1.1",
]
